CLN Moves plot to _plot
