--- sklearn/metrics/plot/__init__.py
@@ -1,6 +1,0 @@
-from .roc_curve import plot_roc_curve, RocCurveVisualizer
-
-__all__ = [
-    'plot_roc_curve',
-    'RocCurveVisualizer'
-]
